Experimental LuaRef lifetime branding
As a side effect, would fix #97, which is the actual driving motivation behind
this change.

Adds compiletest_rs tests for context lifetime invariance, and all the tests and
compiletest_rs tests pass, but the documentation and doctests are *all* screwed
up.

There are some outstanding API design questions, but the strategy SEEMS to
actually work.