Remove the failure crate as a dependency
Originally this was added to make handling complex multi-layer errors while
using the `failure` crate possible.  This commit returns things to the previous
state, where implementing `failure::Fail::cause` is impossible for types that
contain `std::error::Error` and vice versa.  This means that someone using rlua
while also using `failure` in their application will not be able to walk full
cause chains by using `std::error::Error::source`.

This change was made because the situation behind the original change has
shifted: `failure` is no longer heading for 1.0 in its current state, and it is
not currently recommended for library authors to depend on it.  Improvements to
the `std::error::Error` trait that `failure` brought are headed for
stabilization as part of the `Error` trait itself (`source`, which currently
exists, and `backtrace`, which hopefully will be stabilized soon), and hopefully
when this occurs `failure` will return to using std's `Error` trait and the
interoperability pain will go away.  Also, using the `std::error::Error` trait
means that interoperability with error handling strategies *other* thain
`failure` will be much easier.

I apologize again for the API churn.