Two large changes: support for changing _ENV and API change for chunk loading
The second change makes the first change less painful in the API.  I'm not
totally happy with the "builder pattern" of `Chunk`, especially that "loading" a
`Chunk` doesn't actually do anything, but it's not really avoidable.  You cannot
parse lua source until you know whether the user expects it to be an expression
or not (`exec` vs `eval`), and you cannot set the name of a chunk after it is
loaded.  The best I could do is load the code *twice* to ensure on the call to
`load` that it actually compiles, but that is horrifically ineficient.