add `load_from_std_lib` function to `Lua` state
This function allows loading of any portion of the standard libraries
after initial state creation, except debug for unsafety reasons.
Also added is `unsafe_load_from_std_lib` which *does* also allow loading
the debug library.

Specifically, this is filling a need to load the debug library after an
error is encountered so that a traceback can be generated without
risking exposing the debug library anywhere else in the code.